update: 인증 주입 함수 재정의
