+++ hobbyroom/auth/__init__.py
@@ -1,0 +1,3 @@
+from .depends import get_current_user_id
+
+__all__ = ["get_current_user_id"]
